Change BigDecimal to Integer for avoid bug of class type not match in Seata 0.6.1

+++ cloud-seata-mybatis/pay-service/src/main/java/io/github/helloworlde/pay/dao/AccountDao.java
@@ -6,8 +6,6 @@ import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 
-import java.math.BigDecimal;
-
 /**
  * @author HelloWood
  */
@@ -21,7 +19,7 @@ public interface AccountDao {
      * @return 账户余额
      */
     @Select("SELECT balance FROM account WHERE id = #{userId}")
-    BigDecimal getBalance(@Param("userId") Long userId);
+    Integer getBalance(@Param("userId") Long userId);
 
     /**
      * 扣减余额
@@ -30,5 +28,5 @@ public interface AccountDao {
      * @return 影响记录行数
      */
     @Update("UPDATE account SET balance = balance - #{price} WHERE id = 1")
-    Integer reduceBalance(@Param("price") BigDecimal price);
+    Integer reduceBalance(@Param("price") Integer price);
 }
